chore: migrate pure to TS
