--- packages/cli-testing-library/types/event-map.d.ts
@@ -1,12 +1,0 @@
-import {TestInstance} from '../types'
-declare const eventMap: {
-  sigterm: (instance: TestInstance) => Promise<void>
-  sigkill: (instance: TestInstance) => Promise<void>
-  write: (
-    instance: TestInstance,
-    props: {
-      value: string
-    },
-  ) => boolean
-}
-export {eventMap}
